Ecom-admin - kept in monolith architecture for simplicity.
Ecom-app
- broken down into 6 services in total.
- using aggreagate service to call the 5 services.
    - send event message to rabbit MQ for the 5 services to consume.
- CMS
- OMS
    - combined cart, order and order cancel service. Could break down further but unnecessary.
    - ordering is functional but because event driven asynchronous, it's stateless,
      need to hard pass in request parameter of order id and other sensitive infomation.
      will get resolved with security, token implemented.
- changed from imperative to reactive programming.
- all operations like create, update and delete are event driven asynchronous using RabbitMQ.
- only read operations are non-blocking synchrnous.
- security removed for this part
    - passing userID for now
    - will change after discovery server, eureka, then gateway. Auth server.
- services will retain their endpoint/controller for easier diagnostic along the message processing from RabbitMQ.

+++ ECom-app/OMS/src/main/java/com/itsthatjun/ecommerce/mbg/model/OrderReturnApply.java
@@ -14,7 +14,7 @@ public class OrderReturnApply {
 
     private String orderSn;
 
-    private Long memberId;
+    private Integer memberId;
 
     private BigDecimal returnAmount;
 
@@ -94,11 +94,11 @@ public class OrderReturnApply {
         this.orderSn = orderSn;
     }
 
-    public Long getMemberId() {
+    public Integer getMemberId() {
         return memberId;
     }
 
-    public void setMemberId(Long memberId) {
+    public void setMemberId(Integer memberId) {
         this.memberId = memberId;
     }
